Added minimum alignment score threshold

+++ src/options.h
@@ -50,7 +50,7 @@ class Options {
   int32_t flanklen; // used in models
   int32_t regionsize; // region in bam file to search for reads around the STR
   int32_t min_match;  // minimum matching basepairs on each end of enclosing read
-  int32_t min_score;  // minimum alignment score (out of 100)
+  double min_score;  // minimum alignment score (out of 100)
   // Likelihood weights
   double frr_weight;
   double enclosing_weight;
